rename: Filter -> commonFilter로 변경
